backend/camera_service_toyoda/requirements.txt
absl-py==0.9.0
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.10
astor==0.8.1
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
Cython==0.29.21
Django==3.0.8
django-celery==3.3.1
django-cors-headers==3.4.0
djangorestframework==3.11.0
drf-yasg==1.17.1
gast==0.4.0
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
kafka-python==2.0.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
kombu==3.0.37
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.0
memory-profiler==0.58.0
numpy==1.19.1
object-detection==0.1
opencv-contrib-python==4.5.1.48
opencv-python==4.3.0.36
packaging==20.4
pandas==1.1.0
Pillow==7.2.0
protobuf==3.13.0
psutil==5.8.0
pymongo==3.10.1
pyparsing==2.4.7
pypng==0.0.20
PyQRCode==1.2.1
pyserial==3.4
python-dateutil==2.8.1
pytz==2020.1
pyzbar==0.1.8
redis==3.5.3
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
slim==0.1
sqlparse==0.3.1
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
uritemplate==3.0.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
